+++ aeron-client/src/main/c/concurrent/aeron_logbuffer_descriptor.h
@@ -90,7 +90,7 @@ inline int32_t aeron_logbuffer_term_id(int64_t raw_tail)
     return (int32_t)(raw_tail >> 32);
 }
 
-inline int64_t aeron_logbuffer_compute_term_count(int32_t term_id, int32_t initial_term_id)
+inline int32_t aeron_logbuffer_compute_term_count(int32_t term_id, int32_t initial_term_id)
 {
     return aeron_sub_wrap_i32(term_id, initial_term_id);
 }
@@ -102,11 +102,11 @@ inline size_t aeron_logbuffer_index_by_position(int64_t position, size_t positio
 
 inline size_t aeron_logbuffer_index_by_term(int32_t initial_term_id, int32_t active_term_id)
 {
-    int64_t term_count = aeron_logbuffer_compute_term_count(active_term_id, initial_term_id);
+    int32_t term_count = aeron_logbuffer_compute_term_count(active_term_id, initial_term_id);
     return (size_t)(term_count % AERON_LOGBUFFER_PARTITION_COUNT);
 }
 
-inline size_t aeron_logbuffer_index_by_term_count(int64_t term_count)
+inline size_t aeron_logbuffer_index_by_term_count(int32_t term_count)
 {
     return (size_t)(term_count % AERON_LOGBUFFER_PARTITION_COUNT);
 }
